LWRPs fixed to notify only when an action is executed

+++ providers/admin.rb
@@ -13,15 +13,16 @@ action :create do
   db_host = new_resource.db_host || node['postfixadmin']['database']['host']
   ssl = new_resource.ssl || node['postfixadmin']['ssl']
 
-  converge_by("Create #{new_resource}") do
-    db = PostfixAdmin::MySQL.new(db_user, db_password, db_name, db_host)
-    ruby_block "create admin user #{user}" do
-      block do
-        result = PostfixAdmin::API.createAdmin(user, new_resource.password, setup_password, ssl)
-        Chef::Log.info("Created #{new_resource}: #{result}")
+  db = PostfixAdmin::MySQL.new(db_user, db_password, db_name, db_host)
+  unless db.adminExists?(user)
+    converge_by("Create #{new_resource}") do
+      ruby_block "create admin user #{user}" do
+        block do
+          result = PostfixAdmin::API.createAdmin(user, new_resource.password, setup_password, ssl)
+          Chef::Log.info("Created #{new_resource}: #{result}")
+        end
+        action :create
       end
-      not_if do db.adminExists?(user) end
-      action :create
     end
   end
 
@@ -35,15 +36,16 @@ action :remove do
   db_name = new_resource.db_name || node['postfixadmin']['database']['name']
   db_host = new_resource.db_host || node['postfixadmin']['database']['host']
 
-  converge_by("Remove #{new_resource}") do
-    db = PostfixAdmin::MySQL.new(db_user, db_password, db_name, db_host)
-    ruby_block "remove admin user #{user}" do
-      block do
-        deleted = db.removeAdmin(user)
-        Chef::Log.info("Deleted #{new_resource}") if deleted
+  db = PostfixAdmin::MySQL.new(db_user, db_password, db_name, db_host)
+  if db.adminExists?(user)
+    converge_by("Remove #{new_resource}") do
+      ruby_block "remove admin user #{user}" do
+        block do
+          deleted = db.removeAdmin(user)
+          Chef::Log.info("Deleted #{new_resource}") if deleted
+        end
+        action :create
       end
-      only_if do db.adminExists?(user) end
-      action :create
     end
   end
